Rename client and app mainjs files for readability
