+++ src/client/components/NavBar.jsx
@@ -2,24 +2,24 @@ import React, { Component } from 'react';
 
 class NavBar extends Component
 {
-    constructor(props)
-    {
-        super(props);
-    }
+  constructor(props)
+  {
+    super(props);
+  }
 
-    render()
-    {
-        let user = this.props.user;
+  render()
+  {
+    let user = this.props.user;
 
-        return (
-            <nav>
-                <h1>CAH</h1>
-                <p onClick={this.props.loginHandler.bind(null)}>
-                    {user.name ? user.name : 'Name'}
-                </p>
-            </nav>
-        );
-    }
+    return (
+      <nav>
+        <h1>CAH</h1>
+        <p onClick={this.props.loginHandler.bind(null)}>
+          {user.name ? user.name : 'Name'}
+        </p>
+      </nav>
+    );
+  }
 }
 
 export default NavBar;
