+++ src/client/styles/components/footerbar.scss
@@ -1,4 +1,4 @@
-.FooterBar {
+.footerbar {
   width: 30em;
   height: 3em;
   margin-top: 4em;
@@ -9,7 +9,7 @@
   justify-content: space-between;
 }
 
-.timer {
+.footerbar__timer {
   width: 6em;
   height: 3em;
   background-color: #66BB6A;
@@ -21,16 +21,16 @@
     color: white;
     font-size: 21px;
     margin: 0;
-   };
+  };
 }
 
-.users {
+.footerbar__users {
   display: flex;
   flex-direction: row;
   padding: 0 10px;
 }
 
-.users__user {
+.user {
   width: 2em;
   height: 2em;
   background-color: #9C27B0;
